requirements.txt
pandas
numpy
faker
kaggle
python-dotenv
pyspark
azure-storage-blob